Refactor verifyPassword and useCase for actions to useHook
