+++ src/domain/repositories/InstanceRepository.ts
@@ -5,7 +5,6 @@ export interface InstanceRepository {
     getBaseUrl(): string;
     getInstanceVersion(): FutureData<string>;
     getLocales(type: LocaleType): FutureData<Locale[]>;
-    verifyPassword(password: string): FutureData<true>;
 }
 
 export type LocaleType = "dbLocale" | "uiLocale";
